requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
astroid=2.3.3=py37_0
backcall=0.1.0=pypi_0
blas=1.0=mkl
boto3=1.12.45=pypi_0
botocore=1.15.45=pypi_0
ca-certificates=2020.1.1=0
certifi=2020.4.5.1=py37_0
cffi=1.13.2=py37h2e261b9_0
chardet=3.0.4=pypi_0
click=7.1.1=pypi_0
cudatoolkit=10.1.243=h6bb024c_0
cycler=0.10.0=pypi_0
cython=0.29.14=pypi_0
decorator=4.4.1=pypi_0
docutils=0.15.2=pypi_0
fairseq=0.9.0=dev_0
filelock=3.0.12=pypi_0
freetype=2.9.1=h8a8886c_1
future=0.18.2=pypi_0
idna=2.9=pypi_0
imageio=2.6.1=pypi_0
intel-openmp=2020.0=166
ipdb=0.12.3=pypi_0
ipython=7.12.0=pypi_0
ipython-genutils=0.2.0=pypi_0
isort=4.3.21=py37_0
jedi=0.17.0=pypi_0
jmespath=0.9.5=pypi_0
joblib=0.14.1=pypi_0
jpeg=9b=h024ee3a_2
kiwisolver=1.1.0=pypi_0
lazy-object-proxy=1.4.3=py37h7b6447c_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_0
matplotlib=3.1.3=pypi_0
mccabe=0.6.1=pypi_0
mkl=2020.0=166
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.0.15=py37ha843d7b_0
mkl_random=1.1.0=py37hd6b4f25_0
ncurses=6.1=he6710b0_1
networkx=2.4=pypi_0
ninja=1.9.0=py37hfd86e86_0
nltk=3.4.5=pypi_0
numpy=1.18.1=py37h4f9e942_0
numpy-base=1.18.1=py37hde5b4d6_1
olefile=0.46=py37_0
openssl=1.1.1g=h7b6447c_0
parso=0.7.0=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=7.0.0=py37hb39fc2d_0
pip=20.0.2=py37_1
portalocker=1.5.2=pypi_0
prompt-toolkit=3.0.3=pypi_0
protobuf=3.11.3=pypi_0
ptyprocess=0.6.0=pypi_0
pycparser=2.19=py37_0
pygments=2.5.2=pypi_0
pylint=2.4.4=py37_0
pyparsing=2.4.6=pypi_0
python=3.7.6=h0371630_2
python-dateutil=2.8.1=pypi_0
pytorch=1.4.0=py3.7_cuda10.1.243_cudnn7.6.3_0
pywavelets=1.1.1=pypi_0
readline=7.0=h7b6447c_5
regex=2020.1.8=pypi_0
requests=2.23.0=pypi_0
s3transfer=0.3.3=pypi_0
sacrebleu=1.4.3=pypi_0
sacremoses=0.0.41=pypi_0
scikit-image=0.16.2=pypi_0
scipy=1.4.1=pypi_0
sentencepiece=0.1.85=pypi_0
setuptools=45.1.0=py37_0
six=1.14.0=py37_0
sqlite=3.31.1=h7b6447c_0
tensorboardx=2.0=pypi_0
tk=8.6.8=hbc83047_0
tokenizers=0.5.2=pypi_0
torchvision=0.5.0=py37_cu101
tqdm=4.42.1=pypi_0
traitlets=4.3.3=pypi_0
transformers=2.8.0=pypi_0
typed-ast=1.4.1=pypi_0
typing=3.7.4.1=pypi_0
urllib3=1.25.9=pypi_0
wcwidth=0.1.8=pypi_0
wheel=0.34.2=py37_0
wrapt=1.11.2=pypi_0
xz=5.2.4=h14c3975_4
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
